[refactor] JCF, 파일 서비스delete, Basic 서비스 추가
