+++ build.gradle
@@ -1,28 +1,36 @@
 plugins {
-    id 'java'
-    id 'application'
-    id 'org.springframework.boot' version '3.2.0' // ✅ Spring Boot 플러그인 추가
-    id 'io.spring.dependency-management' version '1.1.3' // ✅ 의존성 관리 플러그인
+	id 'java'
+	id 'org.springframework.boot' version '3.4.2'
+	id 'io.spring.dependency-management' version '1.1.7'
 }
 
 group = 'com.sprint.mission'
-version = '1.0-SNAPSHOT'
+version = '0.0.1-SNAPSHOT'
 
-repositories {
-    mavenCentral()
+java {
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(17)
+	}
 }
 
-application {
-    mainClass = 'com.sprint.mission.discodeit.JavaApplication'
+configurations {
+	compileOnly {
+		extendsFrom annotationProcessor
+	}
+}
+
+repositories {
+	mavenCentral()
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter' // ✅ 스프링 핵심 의존성 추가
-    implementation 'org.springframework.boot:spring-boot-starter-web' // ✅ 웹 개발용 (REST API 등)
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
 
-test {
-    useJUnitPlatform()
+tasks.named('test') {
+	useJUnitPlatform()
 }
