+++ settings.gradle
@@ -1,2 +1,1 @@
-rootProject.name = '1-sprint-mission'
-
+rootProject.name = 'discodeit'
