+++ src/main/resources/application.yaml
@@ -1,0 +1,1 @@
+spring.application.name=discodeit
